GraphBG component file + style restructure
