+++ src/modules/common/components/graph-background/graph-background.styles.less
@@ -1,4 +1,6 @@
-#canvas_container {
+@import '~assets/styles/shared.less';
+
+.CanvasContainer {
   height: 100%;
   position: fixed;
   width: 100%;
