+++ src/modules/topics/components/topic-list/topic-list.styles.less
@@ -14,13 +14,6 @@
   &:nth-child(3) {
     margin: 0;
   }
-
-  > a {
-    display: block;
-    text-decoration: none;
-    cursor: pointer;
-  }
-
 }
 
 @media @breakpoint-small-desktop {
